r: modularize `timm` and `torchvision` models.
