+++ models/vision/backbone/timm.py
@@ -1,9 +1,9 @@
 import timm
-import torchvision.models as TorchvisionModels
+
 from utils.models import drop_layers_after
 
 
-def timm_feature_extractor(model_id, *args, **kwargs):
+def timm_feature_extractor(model_id, drop_after=None, reset_classifier=True, *args, **kwargs):
     """
     Load model(and pretrained-weights) implemented in `timm`.
 
@@ -13,32 +13,9 @@ def timm_feature_extractor(model_id, *args, **kwargs):
     ----------
     model_id: str
         Exact name of model to use. We look for `torchvision.models.{model_id}`.
-    kwargs: dict
-        kwargs for building model.
-    Returns
-    -------
-    nn.Module
-        feature_extractor network that can be used in multiple subtasks by plugging in different downstream heads.
-    """
-    # detach final classification head(make it feature extractor)
-    # find model with same id & create model
-    model = timm.create_model(model_id, num_classes=0, *args, **kwargs)
-    return model
-
-
-def torchvision_feature_extractor(model_id, drop_after=None, *args, **kwargs):
-    """
-    Load model(and pretrained-weights) implemented in `torchvision.models`. Although some of our custom
-    architecture implementation is also sort of based on torchvision, we implement this method to support more
-    models and access to pretrained checkpoints.
-
-    Model catalog can be found in: https://pytorch.org/vision/stable/models.html#id1
-    TODO: support `torchvision` models for other tasks. e.g. video.
-
-    Parameters
-    ----------
-    model_id: str
-        Exact name of model to use. We look for `torchvision.models.{model_id}`.
+    reset_classifier: bool
+        Reset classifier automatically drops the classifier without the need to specify `drop_after` for most timm
+        models.
     drop_after: str
         Remove all the layers including and after `drop_after` layer. For example, in the example below of `resnet50`,
         we want to set `drop_after` = "avgpool" to drop the final 2 layers and output feature map.
@@ -52,16 +29,18 @@ def torchvision_feature_extractor(model_id, drop_after=None, *args, **kwargs):
             (fc): Linear(in_features=512, out_features=1000, bias=True)
         )
         ```
-    kwargs: dict
-        kwargs for building model.
     Returns
     -------
     nn.Module
         feature_extractor network that can be used in multiple subtasks by plugging in different downstream heads.
     """
+    # detach final classification head(make it feature extractor)
     # find model with same id & create model
-    model = getattr(TorchvisionModels, str(model_id))(*args, **kwargs)
+    model = timm.create_model(model_id, num_classes=0, *args, **kwargs)
+    if reset_classifier:
+        model.reset_classifier(0, "")
     # detach final classification head(make it feature extractor)
     if drop_after:
         model = drop_layers_after(model, drop_after)
+
     return model
