+++ utils/__init__.py
@@ -1,9 +1,13 @@
 from . import (  # noqa F401
+    bbox,
     callbacks,
     configs,
     experiment,
+    image,
+    logging,
     models,
     optimizers,
+    pretrained,
     verbose,
     visualization,
 )
